refactor(widgets): dedup IDimensions interface
