+++ src/app/widgets/svg-boolean-switch/svg-boolean-switch.component.ts
@@ -1,7 +1,6 @@
 import { Component, DoCheck, input, output, ChangeDetectionStrategy } from '@angular/core';
-import type { IDynamicControl } from '../../core/interfaces/widgets-interface';
+import type { IDynamicControl, IDimensions } from '../../core/interfaces/widgets-interface';
 import type { ITheme } from '../../core/services/app-service';
-import { IDimensions } from '../widget-boolean-switch/widget-boolean-switch.component';
 import { createSwipeGuard } from '../../core/utils/pointer-swipe-guard.util';
 
 @Component({
